add exeptions and finished all crud requests

+++ src/js/script.js
@@ -2,11 +2,11 @@ window.onload = function() {
 	init();
 }
 
-
-
 function init() {
 	addEventListeners();
 }
+
+let count = 0;
 
 function addEventListeners() {
 	const btnsCrudElem = document.querySelectorAll(".main-block__buttons-crud");	
@@ -14,13 +14,13 @@ function addEventListeners() {
 	const btnsInChagesBlockElem = document.querySelectorAll(".changes-block__btn");
 	
 	const chagesBlockElem = document.querySelector(".changes-block");
+	const chagesBlockAllElems = document.querySelectorAll(".changes-block__elem-block");
 
 	let idToUpdate;
-	let count = 0;
-
+	//addEventListener fo buttons with DB
 	for (let i = 0; i < btnsStorageElem.length; i++) {
 		btnsStorageElem[i].addEventListener("click", function(event) {
-			let target = event.target;
+			const target = event.target;
 			
 			for (let j = 0; j < btnsStorageElem.length; j++) {
 				btnsStorageElem[j].classList.remove("active");
@@ -40,10 +40,14 @@ function addEventListeners() {
 			}
 		});
 	}
-	
+	//addEventListener fo buttons CRUD
 	for (let i = 0; i < btnsCrudElem.length; i++) {
 		btnsCrudElem[i].addEventListener("click", function(event) {
-			let target = event.target;
+			if (personDAO === undefined) {
+				alert("Please, select database first");
+				return;
+			}
+			const target = event.target;
 			//read
 			if (target.closest(".btn-read")) {
 				document.querySelector(".main-block__tbody").innerHTML = "";
@@ -56,35 +60,39 @@ function addEventListeners() {
 					addInTable(items);
 				}
 			}
+
+			for (let j = 0; j < chagesBlockAllElems.length; j++) {
+				chagesBlockAllElems[j].classList.remove("active");
+			}
+
 			if (target.closest(".btn-add")) {
 				chagesBlockElem.querySelector(".add").classList.toggle("active");
-				chagesBlockElem.querySelector(".delete").classList.remove("active");
-				chagesBlockElem.querySelector(".update-enter-id").classList.remove("active");
 			}			
 			if (target.closest(".btn-delete")) {
 				chagesBlockElem.querySelector(".delete").classList.toggle("active");
-				chagesBlockElem.querySelector(".add").classList.remove("active");
-				chagesBlockElem.querySelector(".update-enter-id").classList.remove("active");
 			}			
 			if (target.closest(".btn-update")) {
 				chagesBlockElem.querySelector(".update-enter-id").classList.toggle("active");
-				chagesBlockElem.querySelector(".delete").classList.remove("active");
-				chagesBlockElem.querySelector(".add").classList.remove("active");
 			}
 		});
 	}
-
+	//addEventListener fo blocks in which you change DB
 	for (let i = 0; i < btnsInChagesBlockElem.length; i++) {
 		btnsInChagesBlockElem[i].addEventListener("click", function(event) {
-			let target = event.target;
+			const target = event.target;
 			//add
 			if (target.closest(".btn-addintable")) {
 				let parent = target.closest(".btn-addintable").parentElement;
 				let name = parent.querySelector(".name").value;
 				let surname = parent.querySelector(".surname").value;
 				let age = parent.querySelector(".age").value;
+				//exeption if one of the inputs is empty
+				if(name && surname && age === "") {
+					alert("Please, fill in all fields");
+					return;
+				}
 
-				let personObj = getPersonObj(generateId(), name, surname, age);
+				const personObj = getPersonObj(generateId(), name, surname, age);
 				let items = personDAO.addPerson(personObj);
 
 				if (personDAO instanceof PersonDAOIndexedDB) {
@@ -102,8 +110,12 @@ function addEventListeners() {
 			}
 			//delete
 			if (target.closest(".btn-deletefromtable")) {
-				let parent = target.closest(".btn-deletefromtable").parentElement;
-				let id = parent.querySelector(".id-delete").value;
+				const parent = target.closest(".btn-deletefromtable").parentElement;
+				const id = parent.querySelector(".id-delete").value;
+				if(getWarningForId(id) === false) {
+					return; 
+				}
+
 				personDAO.deletePerson(parseInt(id));
 
 				deleteFromTable(id);
@@ -111,26 +123,35 @@ function addEventListeners() {
 			}
 			//enter id of person to update
 			if (target.closest(".btn-next")) {
-				let parent = target.closest(".btn-next").parentElement;
+				const parent = target.closest(".btn-next").parentElement;
 				idToUpdate = parent.querySelector(".id-to-update").value;
-				const tecTr = document.getElementById("tr" + idToUpdate).querySelectorAll("td");
+				if(getWarningForId(idToUpdate) === false) {
+					return; 
+				}
+
+				const currentTr = document.getElementById("tr" + idToUpdate).querySelectorAll("td");
 
 				chagesBlockElem.querySelector(".update").classList.add("active");
 				chagesBlockElem.querySelector(".update-enter-id").classList.remove("active");
 				let arrayInputs = chagesBlockElem.querySelector(".update").querySelectorAll(".form-control-input");
 				
 				for (let i = 0; i < arrayInputs.length; i++) {
-					arrayInputs[i].value = tecTr[i + 1].innerHTML;
+					arrayInputs[i].value = currentTr[i + 1].innerHTML;
 				}
 				parent.querySelector(".id-to-update").value = "";
 			}
 			//update
 			if (target.closest(".btn-updatetable")) {
-				let parent = target.closest(".btn-updatetable").parentElement;
+				const parent = target.closest(".btn-updatetable").parentElement;
 				chagesBlockElem.querySelector(".update").classList.remove("active");
-				let arrayInputs = chagesBlockElem.querySelector(".update").querySelectorAll(".form-control-input");
+				const arrayInputs = chagesBlockElem.querySelector(".update").querySelectorAll(".form-control-input");
+				//exeption if one of the inputs is empty
+				if(arrayInputs[0].value && arrayInputs[1].value && arrayInputs[2].value === "") {
+					alert("Please, fill in all fields");
+					return;
+				}
 
-				let personObj = getPersonObj(parseInt(idToUpdate), arrayInputs[0].value, arrayInputs[1].value, arrayInputs[2].value);
+				const personObj = getPersonObj(parseInt(idToUpdate), arrayInputs[0].value, arrayInputs[1].value, arrayInputs[2].value);
 
 				personDAO.updatePerson(personObj);
 				let currentTr = document.getElementById("tr" + idToUpdate).querySelectorAll("td");
@@ -143,15 +164,42 @@ function addEventListeners() {
 		});
 	}
 }
-let count = 0;
+
+function getWarningForId(id) {
+	//exeption if input is empty
+	if(!!id !== true) {
+		alert("Please, enter id");
+		return false;
+	}
+	//exeption if in IndexedDB missing such id
+	if (personDAO instanceof PersonDAOIndexedDB) {
+		let item = personDAO.getPerson(parseInt(id));
+		item.then(person => {
+			if (person.length === 0) {
+				alert("There is no such id in DB. Please, enter correct id");
+				return false;
+			}
+		});
+	}
+	//exeption if in others DB missing such id
+	console.log(personDAO.getPerson(parseInt(id)))
+	if (personDAO.getPerson(parseInt(id)).length === 0) {
+		alert("There is no such id in DB. Please, enter correct id");
+		return false;
+	}
+}
+
 function generateId() {
 	if (personDAO instanceof PersonDAOLocalStorage) { 
-		count = JSON.parse(localStorage.getItem('PersonList')).length || 0;
+		if(JSON.parse(localStorage.getItem('PersonList'))) {
+			let array = personDAO.readAllPersons();
+			return ++array[array.length - 1].id;
+		}
 	}
 	if (personDAO instanceof PersonDAOIndexedDB) { 
 		return "";
 	}
-	return count++;
+	return ++count;
 }
 
 function getPersonObj(id, name, surname, age) {
